Add Expected
The Expected template class is a wrapper that may contain an object of
the template type T or an error. It is intended to be used as a return
type for operations which may fail. On success, the returned object
contains the expected result. In case of failure, it instead contains an
object that describes the error encountered.

The expected/unexpected idiom provides a useful alternative to
traditional error handling mechanisms such as exceptions and error
codes.

Compared to the use of exceptions, the Expected approach

- can be used in device code and can be safely and easily propagated
across thread boundaries
- allows for inexpensive local handling of the "bad path" when an
operation fails to return the expected value
- improves error visibility in code review by making the possibility of
error explicit in a function's return type

Compared to error codes, Expected objects

- do not monopolize the return channel
- are not easily ignored (if the user wants to retrieve the contained
value)